Split model focus comparison copy
