+++ frontend/app/(localized)/[locale]/(marketing)/models/[slug]/_lib/model-page-focus-vs/creative-pairs.ts
@@ -1,0 +1,120 @@
+import type { FocusVsPair } from './types';
+
+export const CREATIVE_FOCUS_VS_PAIRS: FocusVsPair[] = [
+{
+    slugA: 'ltx-2-fast',
+    slugB: 'ltx-2',
+    nameA: 'LTX 2 Fast',
+    nameB: 'LTX 2',
+    copyA: {
+      en: {
+        title: 'Use LTX 2 Fast when you want:',
+        items: [
+          'High-volume drafts and iteration speed',
+          'Quick concept testing and pacing checks',
+          'Rough cuts before finals',
+        ],
+      },
+      fr: {
+        title: 'Utilisez LTX 2 Fast quand vous voulez :',
+        items: [
+          'Tests en volume et vitesse d’itération',
+          'Tests de concepts rapides et vérification du pacing',
+          'Rough cuts avant les finals',
+        ],
+      },
+      es: {
+        title: 'Usa LTX 2 Fast cuando quieras:',
+        items: [
+          'Borradores en volumen y velocidad de iteración',
+          'Pruebas rápidas de concepto y pacing',
+          'Rough cuts antes de finales',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use LTX 2 when you need:',
+        items: [
+          'Polished client-ready deliverables',
+          'Higher resolution and smoother motion',
+          'Audio-visual sync for finals',
+        ],
+      },
+      fr: {
+        title: 'Utilisez LTX 2 quand vous avez besoin :',
+        items: [
+          'Livrables clients plus soignés',
+          'Résolution plus élevée et mouvement plus fluide',
+          'Sync audio-visuelle pour les finals',
+        ],
+      },
+      es: {
+        title: 'Usa LTX 2 cuando necesites:',
+        items: [
+          'Entregables más pulidos para cliente',
+          'Más resolución y motion más suave',
+          'Sincronía audio-visual para finales',
+        ],
+      },
+    },
+  },
+{
+    slugA: 'nano-banana',
+    slugB: 'nano-banana-pro',
+    nameA: 'Nano Banana',
+    nameB: 'Nano Banana Pro',
+    copyA: {
+      en: {
+        title: 'Use Nano Banana when you want:',
+        items: [
+          'Fast drafts and quick edits',
+          'Rapid concepting and exploration',
+          'Lightweight layout tests',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Nano Banana quand vous voulez :',
+        items: [
+          'Tests rapides et edits rapides',
+          'Concepting rapide et exploration',
+          'Tests de mise en page légers',
+        ],
+      },
+      es: {
+        title: 'Usa Nano Banana cuando quieras:',
+        items: [
+          'Borradores rápidos y ediciones rápidas',
+          'Concepting rápido y exploración',
+          'Pruebas de layout ligeras',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use Nano Banana Pro when you need:',
+        items: [
+          'Clean typography and layouts',
+          'Consistent product families',
+          'High-res finals for campaigns',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Nano Banana Pro quand vous avez besoin :',
+        items: [
+          'Typo et layouts propres',
+          'Familles produit cohérentes',
+          'Finals haute résolution pour campagnes',
+        ],
+      },
+      es: {
+        title: 'Usa Nano Banana Pro cuando necesites:',
+        items: [
+          'Tipografía y layouts limpios',
+          'Familias de producto consistentes',
+          'Finales en alta resolución para campañas',
+        ],
+      },
+    },
+  },
+];
