+++ frontend/app/(localized)/[locale]/(marketing)/models/[slug]/_lib/model-page-focus-vs/foundation-pairs.ts
@@ -1,0 +1,178 @@
+import type { FocusVsPair } from './types';
+
+export const FOUNDATION_FOCUS_VS_PAIRS: FocusVsPair[] = [
+{
+    slugA: 'sora-2',
+    slugB: 'sora-2-pro',
+    nameA: 'Sora 2',
+    nameB: 'Sora 2 Pro',
+    copyA: {
+      en: {
+        title: 'Use Sora 2 when you want:',
+        items: [
+          'Fast idea → clip iteration',
+          'Storyboards, concepts, UGC-style beats, short ads',
+          'A quick first pass where 720p is enough',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Sora 2 quand vous voulez :',
+        items: [
+          'Itération rapide d’idée → clip',
+          'Storyboards, concepts, beats UGC, pubs courtes',
+          'Un premier jet rapide où le 720p suffit',
+        ],
+      },
+      es: {
+        title: 'Usa Sora 2 cuando quieras:',
+        items: [
+          'Iteración rápida de idea → clip',
+          'Storyboards, conceptos, beats estilo UGC, anuncios cortos',
+          'Un primer borrador rápido cuando 720p es suficiente',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use Sora 2 Pro when you need:',
+        items: [
+          'Higher resolution output',
+          'More control for finals (including audio control in the UI)',
+          'Cleaner final takes after you’ve validated the idea',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Sora 2 Pro quand vous avez besoin :',
+        items: [
+          'Sortie en plus haute résolution',
+          'Plus de contrôle pour les finals (y compris le contrôle audio dans l’UI)',
+          'Plans finaux plus propres après validation de l’idée',
+        ],
+      },
+      es: {
+        title: 'Usa Sora 2 Pro cuando necesites:',
+        items: [
+          'Salida de mayor resolución',
+          'Más control para finales (incluye control de audio en la UI)',
+          'Tomas finales más limpias tras validar la idea',
+        ],
+      },
+    },
+  },
+{
+    slugA: 'veo-3-1-fast',
+    slugB: 'veo-3-1',
+    nameA: 'Veo 3.1 Fast',
+    nameB: 'Veo 3.1',
+    copyA: {
+      en: {
+        title: 'Use Veo 3.1 Fast when you want:',
+        items: [
+          'Rapid concept testing and volume drafts',
+          'Cheaper A/B ad variants and social loops',
+          'Quick iteration before upgrading winners',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Veo 3.1 Fast quand vous voulez :',
+        items: [
+          'Tests de concepts rapides et tests en volume',
+          'Variantes A/B moins chères et boucles social',
+          'Itération rapide avant de passer les meilleurs en version supérieure',
+        ],
+      },
+      es: {
+        title: 'Usa Veo 3.1 Fast cuando quieras:',
+        items: [
+          'Pruebas rápidas de concepto y borradores en volumen',
+          'Variantes A/B más baratas y loops sociales',
+          'Iteración rápida antes de pasar a los ganadores',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use Veo 3.1 when you need:',
+        items: [
+          'Higher-fidelity frames and polish',
+          'Sound in the same pass when you want it',
+          'More reliable follow-through on prompts',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Veo 3.1 quand vous avez besoin :',
+        items: [
+          'Fidélité d’image plus élevée et finition',
+          'Son dans le même rendu quand vous en avez besoin',
+          'Suivi des prompts plus fiable',
+        ],
+      },
+      es: {
+        title: 'Usa Veo 3.1 cuando necesites:',
+        items: [
+          'Mayor fidelidad y pulido',
+          'Sonido en la misma pasada cuando lo necesitas',
+          'Seguimiento de prompts más fiable',
+        ],
+      },
+    },
+  },
+{
+    slugA: 'wan-2-5',
+    slugB: 'wan-2-6',
+    nameA: 'Wan 2.5',
+    nameB: 'Wan 2.6',
+    copyA: {
+      en: {
+        title: 'Use Wan 2.5 when you want:',
+        items: [
+          'Native audio in the same render',
+          'Simple short beats at lower cost',
+          'Quick ideation with sound-led timing',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Wan 2.5 quand vous voulez :',
+        items: [
+          'Audio natif dans le même rendu',
+          'Beats courts simples à moindre coût',
+          'Idéation rapide avec timing guidé par le son',
+        ],
+      },
+      es: {
+        title: 'Usa Wan 2.5 cuando quieras:',
+        items: [
+          'Audio nativo en el mismo render',
+          'Beats cortos simples a menor coste',
+          'Ideación rápida con timing guiado por el audio',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use Wan 2.6 when you need:',
+        items: [
+          'Reference-to-video consistency',
+          'Timestamped multi-shot sequences',
+          'More aspect-ratio control and structure',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Wan 2.6 quand vous avez besoin :',
+        items: [
+          'Cohérence en reference-to-video',
+          'Séquences multi-shots horodatées',
+          'Plus de contrôle de format et de structure',
+        ],
+      },
+      es: {
+        title: 'Usa Wan 2.6 cuando necesites:',
+        items: [
+          'Consistencia en reference-to-video',
+          'Secuencias multi-shot con timestamps',
+          'Más control de formato y estructura',
+        ],
+      },
+    },
+  },
+];
