+++ frontend/app/(localized)/[locale]/(marketing)/models/[slug]/_lib/model-page-focus-vs/kling-pairs.ts
@@ -1,0 +1,179 @@
+import type { FocusVsPair } from './types';
+
+export const KLING_FOCUS_VS_PAIRS: FocusVsPair[] = [
+{
+    slugA: 'kling-2-5-turbo',
+    slugB: 'kling-2-6-pro',
+    nameA: 'Kling 2.5 Turbo',
+    nameB: 'Kling 2.6 Pro',
+    copyA: {
+      en: {
+        title: 'Use Kling 2.5 Turbo when you want:',
+        items: [
+          'Fast silent clips with strong motion',
+          'Budget B-roll loops for edits',
+          'Quick look-dev and drafts',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Kling 2.5 Turbo quand vous voulez :',
+        items: [
+          'Clips silencieux rapides avec forte motion',
+          'Boucles B-roll budget pour le montage',
+          'Look-dev rapide et tests',
+        ],
+      },
+      es: {
+        title: 'Usa Kling 2.5 Turbo cuando quieras:',
+        items: [
+          'Clips silenciosos rápidos con motion fuerte',
+          'Loops B-roll económicos para edición',
+          'Look-dev rápido y borradores',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use Kling 2.6 Pro when you need:',
+        items: [
+          'Native audio with dialogue and SFX',
+          'Polished ad/story beats',
+          'Stronger continuity on camera direction',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Kling 2.6 Pro quand vous avez besoin :',
+        items: [
+          'Audio natif avec dialogue et SFX',
+          'Beats pub/story plus polis',
+          'Continuité plus forte sur la direction caméra',
+        ],
+      },
+      es: {
+        title: 'Usa Kling 2.6 Pro cuando necesites:',
+        items: [
+          'Audio nativo con diálogo y SFX',
+          'Beats publicitarios/story más pulidos',
+          'Continuidad más fuerte en la dirección de cámara',
+        ],
+      },
+    },
+  },
+{
+    slugA: 'kling-2-6-pro',
+    slugB: 'kling-3-pro',
+    nameA: 'Kling 2.6 Pro',
+    nameB: 'Kling 3 Pro',
+    copyA: {
+      en: {
+        title: 'Use Kling 2.6 Pro when you want:',
+        items: [
+          'Native audio with dialogue and SFX',
+          'Short cinematic beats without extra setup',
+          'Solid results for 5–10s clips',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Kling 2.6 Pro quand vous voulez :',
+        items: [
+          'Audio natif avec dialogue et SFX',
+          'Beats ciné courts sans setup complexe',
+          'Résultats solides sur 5–10s',
+        ],
+      },
+      es: {
+        title: 'Usa Kling 2.6 Pro cuando quieras:',
+        items: [
+          'Audio nativo con diálogo y SFX',
+          'Beats cinemáticos cortos sin setup extra',
+          'Resultados sólidos en clips de 5–10s',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use Kling 3 Pro when you need:',
+        items: [
+          'Multi-prompt sequencing across scenes',
+          'Element references for stronger continuity',
+          'Voice IDs and shot-type control up to 15s',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Kling 3 Pro quand vous avez besoin :',
+        items: [
+          'Séquençage multi-prompts par scène',
+          'Éléments de référence pour plus de continuité',
+          'Voice IDs et shot-type jusqu’à 15s',
+        ],
+      },
+      es: {
+        title: 'Usa Kling 3 Pro cuando necesites:',
+        items: [
+          'Secuencias multi-prompt por escena',
+          'Referencias de elementos para mayor continuidad',
+          'Voice IDs y control de shot type hasta 15s',
+        ],
+      },
+    },
+  },
+{
+    slugA: 'kling-3-standard',
+    slugB: 'kling-3-pro',
+    nameA: 'Kling 3 Standard',
+    nameB: 'Kling 3 Pro',
+    onlyFor: ['kling-3-standard'],
+    copyA: {
+      en: {
+        title: 'Use Kling 3 Standard when you want:',
+        items: [
+          'Multi-shot control at a lower cost',
+          'Quick ad variants and social promos',
+          'Elements + end frame for consistency',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Kling 3 Standard quand vous voulez :',
+        items: [
+          'Contrôle multi‑shot à moindre coût',
+          'Variantes rapides d’ads et promos social',
+          'Elements + image de fin pour la cohérence',
+        ],
+      },
+      es: {
+        title: 'Usa Kling 3 Standard cuando quieras:',
+        items: [
+          'Control multi‑shot a menor costo',
+          'Variantes rápidas de anuncios y promos sociales',
+          'Elements + end frame para consistencia',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use Kling 3 Pro when you need:',
+        items: [
+          'Shot type control + voice IDs',
+          'More precise coverage for storyboards',
+          'Premium takes and iteration depth',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Kling 3 Pro quand vous avez besoin :',
+        items: [
+          'Shot type + voice IDs',
+          'Couverture plus précise pour storyboard',
+          'Takes premium et itérations plus poussées',
+        ],
+      },
+      es: {
+        title: 'Usa Kling 3 Pro cuando necesites:',
+        items: [
+          'Shot type + voice IDs',
+          'Cobertura más precisa para storyboards',
+          'Tomas premium y más iteración',
+        ],
+      },
+    },
+  },
+];
