+++ frontend/app/(localized)/[locale]/(marketing)/models/[slug]/_lib/model-page-focus-vs/specialized-pairs.ts
@@ -1,0 +1,122 @@
+import type { FocusVsPair } from './types';
+
+export const SPECIALIZED_FOCUS_VS_PAIRS: FocusVsPair[] = [
+{
+    slugA: 'pika-text-to-video',
+    slugB: 'kling-2-5-turbo',
+    nameA: 'Pika 2.2',
+    nameB: 'Kling 2.5 Turbo',
+    onlyFor: ['pika-text-to-video'],
+    copyA: {
+      en: {
+        title: 'Use Pika 2.2 when you want:',
+        items: [
+          'Stylized, social-first motion',
+          'Fast loops and playful variants',
+          'Edit-friendly silent clips',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Pika 2.2 quand vous voulez :',
+        items: [
+          'Motion stylisé, social-first',
+          'Boucles rapides et variantes ludiques',
+          'Clips silencieux faciles à monter',
+        ],
+      },
+      es: {
+        title: 'Usa Pika 2.2 cuando quieras:',
+        items: [
+          'Motion estilizado y social-first',
+          'Loops rápidos y variantes lúdicas',
+          'Clips silenciosos fáciles de editar',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use Kling 2.5 Turbo when you need:',
+        items: [
+          'More cinematic motion and physics',
+          'Camera-forward action beats',
+          'Cleaner realism for product shots',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Kling 2.5 Turbo quand vous avez besoin :',
+        items: [
+          'Motion et physique plus cinématiques',
+          'Beats d’action camera-forward',
+          'Réalisme plus propre pour produits',
+        ],
+      },
+      es: {
+        title: 'Usa Kling 2.5 Turbo cuando necesites:',
+        items: [
+          'Motion y física más cinematográficas',
+          'Beats de acción camera-forward',
+          'Realismo más limpio para productos',
+        ],
+      },
+    },
+  },
+{
+    slugA: 'minimax-hailuo-02-text',
+    slugB: 'ltx-2-fast',
+    nameA: 'Hailuo 02',
+    nameB: 'LTX 2 Fast',
+    onlyFor: ['minimax-hailuo-02-text'],
+    copyA: {
+      en: {
+        title: 'Use Hailuo 02 when you want:',
+        items: [
+          'Physics-heavy motion drafts',
+          'Fast visual iteration on a budget',
+          'Quick storyboard animatics',
+        ],
+      },
+      fr: {
+        title: 'Utilisez Hailuo 02 quand vous voulez :',
+        items: [
+          'Tests motion axés physique',
+          'Itération visuelle rapide à petit budget',
+          'Animatics storyboard rapides',
+        ],
+      },
+      es: {
+        title: 'Usa Hailuo 02 cuando quieras:',
+        items: [
+          'Borradores de motion con física',
+          'Iteración visual rápida con bajo presupuesto',
+          'Animatics de storyboard rápidos',
+        ],
+      },
+    },
+    copyB: {
+      en: {
+        title: 'Use LTX 2 Fast when you need:',
+        items: [
+          'More room per clip for pacing',
+          'Drafts that stay closer to camera intent',
+          'An easy upgrade path to LTX 2.3 Pro',
+        ],
+      },
+      fr: {
+        title: 'Utilisez LTX 2 Fast quand vous avez besoin :',
+        items: [
+          'Plus de durée par clip pour le pacing',
+          'Tests plus proches de l’intention caméra',
+          'Chemin simple vers LTX 2.3 Pro',
+        ],
+      },
+      es: {
+        title: 'Usa LTX 2 Fast cuando necesites:',
+        items: [
+          'Más duración por clip para el pacing',
+          'Borradores más cercanos a la intención de cámara',
+          'Camino fácil hacia LTX 2.3 Pro',
+        ],
+      },
+    },
+  },
+];
